いいね機能修正

+++ app/controllers/favorites_controller.rb
@@ -9,8 +9,10 @@ class FavoritesController < ApplicationController
 
   def destroy
     post = Post.find(params[:post_id])
-    favorite = current_user.favorites.new(post_id: post.id)
-    favorite.destroy
-    redirect_to post_path(post)
+    favorite = current_user.favorites.find_by(post_id: post.id)
+    if favorite
+     favorite.destroy
+    end
+    redirect_to post_path(params[:post_id])
   end
 end
